Restructered everyyhing
