--- data/README.md
@@ -1,13 +1,0 @@
-# Building Mutational Signatures in Cancer using Deep Bayesian Neural Nets
-
-## Data
-
-The `data` folder is intended for storing datasets or data files used in the project.
-
-### `COSMIC_v3.3.1_SBS_GRCh37.txt`
-
-Mutational signatures based on an analysis of 10,952 exomes and 1,048 whole-genomes across 40 distinct types of human cancer. These analyses are based on curated data that were generated by The Cancer Genome Atlas (TCGA), the International Cancer Genome Consortium (ICGC), and a large set of freely available somatic mutations published in peer-reviewed journals. Tab seperated file.
-
-### `WGS_PCAWG.96.csv`
-
-File contains all of the following the pyrimidine single nucleotide variants, N[{C > A, G, or T} or {T > A, G, or C}]N. 4 possible starting nucleotides x 6 pyrimidine variants x 4 ending nucleotides = 96 total combinations. Comma seperated.
