--- documents/README.md
@@ -1,13 +1,0 @@
-# Building Mutational Signatures in Cancer using Deep Bayesian Neural Nets
-
-## Documents
-
-The `documents` folder contains important documents related to school, research, or any relevant materials. These documents may include academic papers, project proposals, and other resources that contribute to the understanding and development of the project.
-
-### `project_proposal.pdf`
-
-Is my project proposal for this project/research.
-
-### `PvA.pdf`
-
-Is my plan of approach for this project/research.
