+++ test/files/test.vcf
@@ -1,0 +1,8 @@
+Thy-AdenoCa	PTC-88C	ICGC:THCA-SA	GRCh37	SNP	2	96521744	96521744	T	C	SOMATIC; dbSNP v147
+Thy-AdenoCa	PTC-88C	ICGC:THCA-SA	GRCh37	SNP	4	96525794	96525794	C	A	SOMATIC; dbSNP v147
+Thy-AdenoCa	PTC-88C	ICGC:THCA-SA	GRCh37	SNP	2	96617125	96617125	T	C	SOMATIC; dbSNP v147
+Thy-AdenoCa	PTC-88C	ICGC:THCA-SA	GRCh37	SNP	2	97817617	97817617	T	C	SOMATIC; dbSNP v147
+Thy-AdenoCa	PTC-88C	ICGC:THCA-SA	GRCh37	SNP	2	97817647	97817647	A	G	SOMATIC; dbSNP v147
+Thy-AdenoCa	PTC-88C	ICGC:THCA-SA	GRCh37	SNP	2	97817659	97817659	C	T	SOMATIC; dbSNP v147
+Thy-AdenoCa	PTC-88C	ICGC:THCA-SA	GRCh37	SNP	2	97817661	97817661	C	T	SOMATIC; dbSNP v147; 1000 genomes;AF=0.033147
+Thy-AdenoCa	PTC-88C	ICGC:THCA-SA	GRCh37	SNP	2	97911732	97911732	A	G	SOMATIC; dbSNP v147
